Add files via upload
The VME is a robust method when we do not need to decompose whole the intrinsic mode functions of the signal. Indeed, if we aim to achieve a particular mode of the signal VME is the best method in this regard (just with knowing an approximation of the frequency band of that specific mode of interest).